feat(bitbucket): convert utils to ts
